Make special "documentation" packages
This lets other camke files find the installed documentation
directories. I needed this because, for example, the tiva and x64
versions of a library may both have different documentation, but if we
are not cross compiling we cannot find the the tiva verison.

So now, to find documentation simply
find_package(package_name_doxygen)
find_package(package_name_sphinx_cmake)